Incorporated the bls library into the project
The repository https://github.com/OliverSherouse/bls is on the PyPI, and it does essentially all that I need at the moment (I've submitted a pull request to the repository that extends its functionality, but it isn't critical). This update uses the bls package to acquire and store the BLS data, and updates the bls query script correspondingly.